migrated SavedTenantApiTest information to AccountApiTest since they're under the same route
